Switched to file based components for the most part
